Move Zero related functionality to a new file
