+++ sdk/shared/src/main/scala/org/ergoplatform/sdk/ContractTemplate.scala
@@ -6,6 +6,7 @@ import io.circe._
 import io.circe.syntax.{EncoderOps, _}
 import org.ergoplatform.ErgoBox
 import org.ergoplatform.sdk.utils.SerializationUtils.{parseString, serializeString}
+import org.ergoplatform.sdk.utils.Zero
 import scalan.RType
 import scalan.RType._
 import scorex.crypto.authds.avltree.batch.BatchAVLProver
@@ -387,124 +388,4 @@ object ContractTemplate {
       }
     })
   }
-}
-
-private trait Zero[T] {
-  def zero: T
-}
-
-private case class CZero[T](zero: T) extends Zero[T]
-
-private trait ZeroLowPriority {
-  implicit def collIsZero[T: Zero: RType]: Zero[Coll[T]] = CZero(Colls.emptyColl[T])
-  implicit def optionIsZero[T: Zero]: Zero[Option[T]] = CZero(None)
-  implicit def pairIsZero[A: Zero, B: Zero]: Zero[(A,B)] = CZero(Zero[A].zero, Zero[B].zero)
-  implicit def funcIsZero[A: Zero, B: Zero]: Zero[A =>B] = CZero((_ : A) => { Zero[B].zero })
-}
-private object Zero extends ZeroLowPriority {
-  def apply[T](implicit z: Zero[T]): Zero[T] = z
-  def zeroOf[T: Zero]: T = Zero[T].zero
-
-  implicit val BooleanIsZero: Zero[Boolean] = CZero(false)
-  implicit val ByteIsZero: Zero[Byte] = CZero(0.toByte)
-  implicit val ShortIsZero: Zero[Short] = CZero(0.toShort)
-  implicit val IntIsZero: Zero[Int] = CZero(0)
-  implicit val LongIsZero: Zero[Long] = CZero(0L)
-  implicit val BigIntIsZero: Zero[BigInt] = CZero(CBigInt(BigInteger.ZERO))
-  implicit val GroupElementIsZero: Zero[GroupElement] = CZero(CGroupElement(CryptoConstants.dlogGroup.identity))
-  implicit val AvlTreeIsZero: Zero[AvlTree] = CZero({
-    val avlProver = new BatchAVLProver[Digest32, Blake2b256.type](keyLength = 32, None)
-    val digest = avlProver.digest
-    val treeData = new AvlTreeData(digest, AvlTreeFlags.AllOperationsAllowed, 32, None)
-    CAvlTree(treeData)
-  })
-  implicit val sigmaPropIsZero: Zero[SigmaProp] = CZero(CSigmaProp(TrivialProp.FalseProp))
-  implicit val AnyIsZero: Zero[Any] = CZero(0)
-  implicit val UnitIsZero: Zero[Unit] = CZero(())
-  implicit val SigmaDslBuilderIsZero: Zero[SigmaDslBuilder] = CZero(CostingSigmaDslBuilder)
-
-  def typeToZero[T](t: RType[T]): Zero[T] = (t match {
-    case BooleanType => Zero[Boolean]
-    case ByteType => Zero[Byte]
-    case ShortType => Zero[Short]
-    case IntType => Zero[Int]
-    case LongType => Zero[Long]
-    case AnyType => Zero[Any]
-    case UnitType => Zero[Unit]
-    case BigIntRType => Zero[BigInt]
-    case BoxRType => Zero[Box]
-    case ContextRType => Zero[Context]
-    case SigmaDslBuilderRType => Zero[SigmaDslBuilder]
-    case HeaderRType => Zero[Header]
-    case PreHeaderRType => Zero[PreHeader]
-    case GroupElementRType => Zero[GroupElement]
-    case AvlTreeRType => Zero[AvlTree]
-    case SigmaPropRType => sigmaPropIsZero
-    case ct: CollType[a] => collIsZero(typeToZero(ct.tItem), ct.tItem)
-    case ct: OptionType[a] => optionIsZero(typeToZero(ct.tA))
-    case ct: PairType[a, b] => pairIsZero(typeToZero(ct.tFst), typeToZero(ct.tSnd))
-    case tt: TupleType => CZero(tt.emptyArray)
-    case ft: FuncType[a, b] => funcIsZero(typeToZero(ft.tDom), typeToZero(ft.tRange))
-    case _ => sys.error(s"Don't know how to compute Zero for type $t")
-  }).asInstanceOf[Zero[T]]
-
-  private val syntheticBox = new ErgoBox(
-    LongIsZero.zero,
-    new ErgoTree(
-      ByteIsZero.zero,
-      IndexedSeq.empty,
-      Right(sigmaPropIsZero.zero)
-    ),
-    Colls.emptyColl,
-    Map.empty,
-    ModifierId @@ ("synthetic_transaction_id"),
-    ShortIsZero.zero,
-    IntIsZero.zero
-  )
-  private val syntheticPreHeader = CPreHeader(
-    ByteIsZero.zero,
-    Colls.emptyColl[Byte],
-    LongIsZero.zero,
-    LongIsZero.zero,
-    IntIsZero.zero,
-    GroupElementIsZero.zero,
-    Colls.emptyColl[Byte]
-  )
-  private val syntheticHeader = CHeader(
-    Colls.emptyColl[Byte],
-    ByteIsZero.zero,
-    Colls.emptyColl[Byte],
-    Colls.emptyColl[Byte],
-    CAvlTree(AvlTreeIsZero.zero),
-    Colls.emptyColl[Byte],
-    timestamp = LongIsZero.zero,
-    nBits = LongIsZero.zero,
-    height = IntIsZero.zero,
-    extensionRoot = Colls.emptyColl[Byte],
-    minerPk = GroupElementIsZero.zero,
-    powOnetimePk = GroupElementIsZero.zero,
-    powNonce = Colls.emptyColl[Byte],
-    powDistance = BigIntIsZero.zero,
-    votes = Colls.emptyColl[Byte]
-  )
-  private val syntheticContext = CostingDataContext(
-    _dataInputs = Colls.emptyColl,
-    headers = Colls.emptyColl,
-    preHeader = syntheticPreHeader,
-    inputs = Colls.emptyColl,
-    outputs = Colls.emptyColl,
-    height = IntIsZero.zero,
-    selfBox = CostingBox(syntheticBox),
-    selfIndex = IntIsZero.zero,
-    lastBlockUtxoRootHash = AvlTreeIsZero.zero,
-    _minerPubKey = Colls.emptyColl[Byte],
-    vars = Colls.emptyColl,
-    activatedScriptVersion = ByteIsZero.zero,
-    currentErgoTreeVersion = ByteIsZero.zero
-  )
-
-  implicit val BoxIsZero: Zero[Box] = CZero(syntheticBox)
-  implicit val ContextIsZero: Zero[Context] = CZero(syntheticContext)
-  implicit val HeaderIsZero: Zero[Header] = CZero(syntheticHeader)
-  implicit val PreHeaderIsZero: Zero[PreHeader] = CZero(syntheticPreHeader)
 }
